+++ firmware/prototyper/src/platform/report.rs
@@ -84,6 +84,12 @@ fn log_reset(board: &BoardInfo) {
             address.get(),
             controller.start().as_usize()
         );
+    } else if let Some(registers) = board.allwinner_f101_watchdog {
+        info!(
+            "{:<30}: Available (F101 watchdog @ 0x{:x})",
+            "Platform Reset Extension",
+            registers.start().as_usize(),
+        );
     } else if board.syscon_poweroff.is_some() || board.syscon_reboot.is_some() {
         info!(
             "{:<30}: Available (syscon: poweroff={}, reboot={})",